refactor(cell): split combat.rs respawn fork into its own file (#244)
combat.rs was 774 lines (over the 700-line cap). The respawn fork
(same-world reanchor vs. cross-world gate-travel) and respawn-target
resolver together dominated the file. Split:

  combat/mod.rs    — dispatch (USE_ABILITY, USE_ABILITY_ON_GROUND,
                     CALL_FOR_AID, RESPAWN, UNSTUCK, RESET_MY_ABILITIES)
  combat/respawn.rs — handle_respawn (same/cross-world fork) +
                      resolve_respawn_target (id → world fallback chain)
  combat/tests.rs   — existing test suite, unchanged behavior

Public surface unchanged: parent declares `pub mod combat;`, and only
`dispatch` is referenced externally.

All 7 combat tests pass; clippy clean; cargo fmt applied.
