+++ crates/services/src/cell/cell_methods/player/combat/mod.rs
@@ -1,0 +1,167 @@
+//! Combat-related player method dispatch — `useAbility`,
+//! `useAbilityOnGroundTarget`, `callForAid` (Defeat Window respawn),
+//! the auto-`respawn` path, and a couple of unimplemented stubs.
+//!
+//! The respawn fork (same-world in-place reanchor vs. cross-world
+//! gate-travel) lives in [`respawn`] so this match stays a thin
+//! "which entry point" dispatch.
+
+use crate::cell::messages::CellToBaseMsg;
+use crate::cell::space_manager::SpaceManager;
+use cimmeria_content_engine::chain::ChainEngine;
+use cimmeria_entity::stats::HEALTH;
+use tokio::sync::mpsc;
+
+use super::constants::*;
+
+mod respawn;
+
+#[cfg(test)]
+mod tests;
+
+pub async fn dispatch(
+    entity_id: u32,
+    method_index: u16,
+    args: &[u8],
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+    engine: &ChainEngine,
+) -> bool {
+    match method_index {
+        CALL_FOR_AID => {
+            if args.len() >= 4 {
+                let respawner_id = i32::from_le_bytes([args[0], args[1], args[2], args[3]]);
+                tracing::info!(entity_id, respawner_id, "callForAid");
+                respawn::handle_respawn(entity_id, respawner_id, tx, space_mgr).await;
+            }
+            true
+        }
+
+        USE_ABILITY => {
+            if args.len() >= 8 {
+                let ability_id = i32::from_le_bytes([args[0], args[1], args[2], args[3]]);
+                let target_id = i32::from_le_bytes([args[4], args[5], args[6], args[7]]);
+                tracing::debug!(entity_id, ability_id, target_id, "useAbility");
+
+                // Snapshot whether the target was alive *before* the ability
+                // resolves. Without this, hitting an already-dead corpse would
+                // re-fire fire_entity_death (and stomp the AI cleanup that
+                // handle_use_ability already performed on the original kill),
+                // double-counting mission progress on every post-death swing.
+                let was_alive_before = if target_id > 0 {
+                    space_mgr.get_entity(target_id as u32).is_some_and(|t| {
+                        !t.is_player && t.stats.get(HEALTH).is_some_and(|s| s.cur > 0)
+                    })
+                } else {
+                    false
+                };
+
+                crate::cell::abilities::handle_use_ability(
+                    entity_id, ability_id, target_id, tx, space_mgr,
+                )
+                .await;
+
+                // Only react to alive→dead transitions caused by *this* call.
+                // handle_use_ability already handles AI/loot/XP on the kill
+                // itself; we only need to fire the content-engine death event
+                // here, since that's a separate concern wired off the killing
+                // player's player_id.
+                if was_alive_before {
+                    let target_eid = target_id as u32;
+                    let just_died = space_mgr
+                        .get_entity(target_eid)
+                        .is_some_and(|t| t.stats.get(HEALTH).is_some_and(|s| s.cur <= 0));
+                    if just_died {
+                        let tag = space_mgr.get_entity(target_eid).and_then(|t| t.tag.clone());
+                        if let Some(tag) = tag {
+                            match space_mgr.get_entity(entity_id).and_then(|e| e.player_id) {
+                                Some(player_id) => {
+                                    crate::cell::content::fire_entity_death(
+                                        entity_id, player_id, &tag, engine, tx, space_mgr,
+                                    )
+                                    .await;
+                                }
+                                None => {
+                                    tracing::warn!(
+                                        entity_id, npc_tag = %tag,
+                                        "Skipping entity_death event: killer entity has no player_id"
+                                    );
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            true
+        }
+
+        USE_ABILITY_ON_GROUND => {
+            if args.len() >= 16 {
+                let ability_id = i32::from_le_bytes([args[0], args[1], args[2], args[3]]);
+                let x = f32::from_le_bytes([args[4], args[5], args[6], args[7]]);
+                let y = f32::from_le_bytes([args[8], args[9], args[10], args[11]]);
+                let z = f32::from_le_bytes([args[12], args[13], args[14], args[15]]);
+                tracing::debug!(entity_id, ability_id, x, y, z, "useAbilityOnGroundTarget");
+
+                // handle_use_ability_on_ground returns the entity IDs of every
+                // NPC that died during this cast (primary + AoE secondaries).
+                // We fire the content-engine death event for each, so kill-
+                // count missions and other death-triggered chains advance for
+                // every AoE kill — not just the primary. Empty Vec means
+                // either no targets in radius, primary cast rejected, or
+                // nothing died.
+                let deaths = crate::cell::abilities::handle_use_ability_on_ground(
+                    entity_id,
+                    ability_id,
+                    [x, y, z],
+                    tx,
+                    space_mgr,
+                )
+                .await;
+
+                if !deaths.is_empty() {
+                    // Resolve player_id once — it doesn't change across kills.
+                    let player_id = space_mgr.get_entity(entity_id).and_then(|e| e.player_id);
+                    for dead_eid in deaths {
+                        let tag = space_mgr.get_entity(dead_eid).and_then(|t| t.tag.clone());
+                        if let Some(tag) = tag {
+                            match player_id {
+                                Some(pid) => {
+                                    crate::cell::content::fire_entity_death(
+                                        entity_id, pid, &tag, engine, tx, space_mgr,
+                                    )
+                                    .await;
+                                }
+                                None => {
+                                    tracing::warn!(
+                                        entity_id, npc_tag = %tag, dead_eid,
+                                        "Skipping entity_death event (ground target): killer entity has no player_id"
+                                    );
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            true
+        }
+
+        RESPAWN => {
+            tracing::debug!(entity_id, "respawn (auto)");
+            respawn::handle_respawn(entity_id, -1, tx, space_mgr).await;
+            true
+        }
+
+        UNSTUCK => {
+            tracing::info!(entity_id, "UNIMPLEMENTED: unstuck");
+            true
+        }
+
+        RESET_MY_ABILITIES => {
+            tracing::info!(entity_id, "UNIMPLEMENTED: resetMyAbilities");
+            true
+        }
+
+        _ => false,
+    }
+}
