+++ crates/services/src/cell/cell_methods/player/combat/respawn.rs
@@ -1,0 +1,240 @@
+//! Respawn fork — same-world in-place reanchor (preserves the cell entity
+//! and the client-side kismet state) vs. cross-world gate-travel (a
+//! genuine space change, instance teardown is unavoidable).
+//!
+//! `resolve_respawn_target` resolves `(world, position)` from the
+//! Defeat-Window-supplied respawner id, falling back through the player's
+//! current world to the Castle default.
+
+use cimmeria_entity::stats::{FOCUS, HEALTH};
+use tokio::sync::mpsc;
+
+use crate::cell::messages::CellToBaseMsg;
+use crate::cell::space_manager::SpaceManager;
+
+/// In-place respawn: keep the cell entity (and instance) alive, send a
+/// targeted client-side burst that re-creates just the pawn actor and
+/// repopulates its visible properties.
+///
+/// Why in-place: `RESET_ENTITIES` (the gate-travel sledgehammer) tears down
+/// every client-side entity, which fires kismet `OnInit`/`OnSpawn` events
+/// and resets door states / completed encounters / triggered sequences.
+/// The user dying in a room with an opened door comes back to find it
+/// closed again. The server-side instance survives (cell entity preserved),
+/// but the *client-side* kismet state was the visible regression.
+///
+/// What we send instead (handled in [`reanchor_player.rs`](../../../base/world_entry/reanchor_player.rs)):
+/// 1. `BASEMSG_CREATE_BASE_PLAYER` — the load-bearing pawn-recreate
+///    primitive. Invokes the client's `createBasePlayer` callback, which
+///    destroys the ragdolled pawn actor and instantiates a fresh standing
+///    one (same hook used on initial login).
+/// 2. `BASEMSG_SPACE_VIEWPORT_INFO` + `BASEMSG_CREATE_CELL_PLAYER` +
+///    `BASEMSG_FORCED_POSITION` — the gate-travel "enter-world body",
+///    keeps the client's space/viewport tables consistent with the new
+///    pawn and snaps it to spawn.
+/// 3. `BeingAppearance` + `onEntityTint` (separate bundle) — replays the
+///    cached appearance args from initial world entry so the new pawn
+///    isn't blank.
+///
+/// AoI entities, kismet sequence state, and the level itself are
+/// untouched because we never send `RESET_ENTITIES`.
+///
+/// Why not the prior in-place attempts (recorded so future-us doesn't
+/// repeat the iteration):
+/// - `onSequence Entity_Spawn` (event 5000): the cooked
+///   `KIS-abilities_human.Death` package's `SeqEvent_EntitySpawn` node
+///   has no output wired to `APawn::TermRagdoll` (confirmed via Ghidra),
+///   so the kismet path is a no-op.
+/// - `CREATE_CELL_PLAYER`-only burst (no `CREATE_BASE_PLAYER` prefix):
+///   instance preserved, but pawn stayed ragdolled. The client's
+///   `createCellPlayer` handler treats a re-issue for an existing
+///   player id as a space/viewport update, not a pawn recreate.
+/// - `CREATE_BASE_PLAYER` prefix without property replay: pawn was
+///   re-created cleanly (un-ragdolled) but had no properties, leaving
+///   the player invisible. Pulling appearance/tint from the cached
+///   world-entry args closed the loop.
+/// - Full `RESET_ENTITIES` reload: clears ragdoll cleanly but resets
+///   all client kismet (door re-closes, encounters re-fire). Rejected.
+///
+/// Cross-world respawn (different world from where the player died) still
+/// falls through to `GateTravel` because the player is genuinely leaving
+/// the space. Instance teardown is unavoidable in that case.
+pub(super) async fn handle_respawn(
+    entity_id: u32,
+    respawner_id: i32,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let space_id = match space_mgr.get_entity(entity_id) {
+        Some(e) => e.space_id.0 as u32,
+        None => {
+            tracing::warn!(entity_id, "respawn: entity not found");
+            return;
+        }
+    };
+
+    let (target_world, spawn_pos) = resolve_respawn_target(respawner_id, entity_id, space_mgr);
+
+    let current_world = space_mgr.get_entity_world_name(entity_id);
+    let same_world = current_world.as_deref() == Some(target_world.as_str());
+
+    // Close the Defeat Window first.
+    let _ = tx
+        .send(CellToBaseMsg::EntityMethodCall {
+            entity_id,
+            method_index: crate::mercury::method_idx::ON_END_AID_WAIT,
+            args: Vec::new(),
+        })
+        .await;
+
+    if !same_world {
+        // Cross-world: full gate-travel reload (player is leaving the space).
+        // Entity existence guaranteed by the early-return above.
+        let entity = space_mgr
+            .get_entity_mut(entity_id)
+            .expect("entity existence checked above");
+        if let Some(player_id) = entity.player_id {
+            super::super::super::inventory::flush_dirty_bandolier_ammo(entity, player_id, tx).await;
+        }
+        space_mgr.destroy_entity(entity_id);
+        tracing::info!(
+            entity_id,
+            from = ?current_world,
+            to = %target_world,
+            "Respawn: cross-world via GateTravel"
+        );
+        let _ = tx
+            .send(CellToBaseMsg::GateTravel {
+                entity_id,
+                target_world_name: target_world,
+                position: spawn_pos,
+                rotation: [0.0; 3],
+            })
+            .await;
+        return;
+    }
+
+    // Same-world: reset cell-entity state in place, then dispatch
+    // `ReanchorPlayer` to drive the client-side pawn recreate +
+    // appearance replay.
+    let entity = space_mgr
+        .get_entity_mut(entity_id)
+        .expect("entity existence checked above");
+    if let Some(h) = entity.stats.get_mut(HEALTH) {
+        h.set_current(h.max);
+    }
+    if let Some(f) = entity.stats.get_mut(FOCUS) {
+        f.set_current(f.max);
+    }
+    let stat_update = entity.stats.serialize_dirty();
+    entity.stats.clear_dirty();
+
+    // Hard-reset state flags + their refcounts. A raw `state_field = 0`
+    // would clear the bits but leave stale counters, which the next
+    // ref-counted unset would interpret as still-positive.
+    entity.clear_all_state_flags();
+    entity.abilities.clear_all_cooldowns();
+
+    space_mgr.update_entity_position(entity_id, spawn_pos, [0, 0, 0], [0.0; 3]);
+
+    // Push the refreshed HEALTH/FOCUS to the HUD via onStatUpdate.
+    if !stat_update.is_empty() {
+        crate::cell::abilities::send_entity_method(
+            entity_id,
+            crate::mercury::method_idx::ON_STAT_UPDATE,
+            stat_update,
+            tx,
+            space_mgr,
+        )
+        .await;
+    }
+
+    // Clear `state_field` on the owning client (lifts BSF_Dead /
+    // BSF_MovementLock / dead-cursor visuals).
+    crate::cell::abilities::send_entity_method(
+        entity_id,
+        crate::mercury::method_idx::ON_STATE_FIELD_UPDATE,
+        0u32.to_le_bytes().to_vec(),
+        tx,
+        space_mgr,
+    )
+    .await;
+
+    // Re-anchor: CREATE_BASE_PLAYER + VIEWPORT + CREATE_CELL_PLAYER +
+    // FORCED_POSITION + cached BeingAppearance/onEntityTint replay.
+    // CREATE_BASE_PLAYER is the load-bearing piece — it triggers the
+    // client's pawn-recreate hook, dropping the ragdoll state.
+    let _ = tx
+        .send(CellToBaseMsg::ReanchorPlayer {
+            entity_id,
+            space_id,
+            position: spawn_pos,
+            rotation: [0.0; 3],
+        })
+        .await;
+}
+
+/// Resolve `(world, position)` for the respawn target.
+///
+/// Priority:
+///   1. Explicit `respawner_id` from the Defeat Window (must be > 0).
+///   2. First respawner registered for the player's current world.
+///   3. Castle default for `Castle_CellBlock` / unknown world.
+///   4. In-place at the player's current position for any other world
+///      (avoids silently teleporting players cross-world).
+///
+/// Operational note: in-place respawn outside Castle can produce death
+/// loops if the player died standing in damaging geometry (lava tile, AoE
+/// pool) and no respawner is configured for that world — they'll respawn
+/// at full health, take the geometry damage tick, and die again. The
+/// clean fix is content-side: every world should ship at least one
+/// respawner. The fallback warn log is the operator signal.
+pub(super) fn resolve_respawn_target(
+    respawner_id: i32,
+    entity_id: u32,
+    space_mgr: &SpaceManager,
+) -> (String, [f32; 3]) {
+    const CASTLE_WORLD: &str = "Castle_CellBlock";
+    const CASTLE_DEFAULT_POS: [f32; 3] = [-334.231, 73.472, -228.026];
+
+    if respawner_id > 0 {
+        if let Some(r) = space_mgr
+            .respawners
+            .iter()
+            .find(|r| r.respawner_id == respawner_id)
+        {
+            return (r.world_name.clone(), r.pos);
+        }
+        tracing::warn!(
+            entity_id,
+            respawner_id,
+            "Respawner not found, falling back to world default"
+        );
+    }
+
+    let world_name = space_mgr.get_entity_world_name(entity_id);
+    if let Some(ref wn) = world_name {
+        if let Some(r) = space_mgr.respawners.iter().find(|r| r.world_name == *wn) {
+            return (r.world_name.clone(), r.pos);
+        }
+    }
+
+    match world_name.as_deref() {
+        Some(CASTLE_WORLD) | None => {
+            tracing::debug!(entity_id, world = ?world_name, "No respawner; using Castle default position");
+            (CASTLE_WORLD.to_string(), CASTLE_DEFAULT_POS)
+        }
+        Some(world) => {
+            let in_place = space_mgr
+                .get_entity(entity_id)
+                .map(|e| [e.position.x, e.position.y, e.position.z])
+                .unwrap_or(CASTLE_DEFAULT_POS);
+            tracing::warn!(
+                entity_id,
+                world = world,
+                "No respawner configured for this world — respawning in place at current position"
+            );
+            (world.to_string(), in_place)
+        }
+    }
+}
